+++ Editor/Importers/YarnProjectImporter.cs
@@ -15,8 +15,10 @@ using System.Collections;
 namespace Yarn.Unity.Editor
 {
     [ScriptedImporter(2, new[] { "yarnproject" }, 1), HelpURL("https://yarnspinner.dev/docs/unity/components/yarn-programs/")]
-    public class YarnProjectImporter : ScriptedImporter
+    [InitializeOnLoad]
+    public class YarnProjectImporter : ScriptedImporter, IYarnErrorSource
     {
+        static YarnProjectImporter() => YarnPreventPlayMode.AddYarnErrorSourceType<YarnProjectImporter>("t:YarnProject");
 
         [System.Serializable]
         public class SerializedDeclaration
@@ -112,11 +114,17 @@ namespace Yarn.Unity.Editor
 
         public bool useAddressableAssets;
 
+        IList<string> IYarnErrorSource.CompileErrors => compileErrors;
+
+        bool IYarnErrorSource.Destroyed => this == null;
+
         public override void OnImportAsset(AssetImportContext ctx)
         {
 #if YARNSPINNER_DEBUG
             UnityEngine.Profiling.Profiler.enabled = true;
 #endif
+
+            YarnPreventPlayMode.AddYarnErrorSource(this);
 
             var project = ScriptableObject.CreateInstance<YarnProject>();
 
@@ -235,7 +243,7 @@ namespace Yarn.Unity.Editor
 
                 return;
             }
-            
+
             if (compilationResult.Program == null)
             {
                 ctx.LogImportError("Internal error: Failed to compile: resulting program was null, but compiler did not report errors.");
@@ -342,7 +350,7 @@ namespace Yarn.Unity.Editor
                             // YarnProjectUtility.UpdateAssetAddresses to
                             // ensure that the appropriate assets have the
                             // appropriate addresses.)
-                            newLocalization.UsesAddressableAssets = true;                            
+                            newLocalization.UsesAddressableAssets = true;
                         }
                         else
                         {
@@ -521,7 +529,8 @@ namespace Yarn.Unity.Editor
 
             var errors = compilationResult.Diagnostics.Where(d => d.Severity == Diagnostic.DiagnosticSeverity.Error);
 
-            if (errors.Count() > 0) {
+            if (errors.Count() > 0)
+            {
                 Debug.LogError($"Can't generate a strings table from a Yarn Project that contains compile errors", null);
                 return null;
             }
@@ -538,8 +547,6 @@ namespace Yarn.Unity.Editor
             });
 
             return stringTableEntries;
-
-
         }
     }