+++ Editor/Utility/YarnEditorUtility.cs
@@ -1,6 +1,8 @@
 using UnityEngine;
 using UnityEditor;
 using System.IO;
+using System.Collections.Generic;
+using System.Linq;
 
 namespace Yarn.Unity.Editor
 {
@@ -212,5 +214,19 @@ namespace Yarn.Unity.Editor
                 ProjectWindowUtil.ShowCreatedAsset(o);
             }
         }
+
+        /// <summary>
+        /// Get all assets of a given type.
+        /// </summary>
+        /// <typeparam name="T">AssetImporter type to search for. Should be convertible from AssetImporter.</typeparam>
+        /// <param name="filterQuery">Asset query (see <see cref="AssetDatabase.FindAssets(string)"/> documentation for formatting).</param>
+        /// <param name="converter">Custom type caster.</param>
+        /// <returns>Enumerable of all assets of a given type.</returns>
+        public static IEnumerable<T> GetAllAssetsOf<T>(string filterQuery, System.Func<AssetImporter, T> converter = null) where T : class
+            => AssetDatabase.FindAssets(filterQuery)
+                .Select(AssetDatabase.GUIDToAssetPath)
+                .Select(AssetImporter.GetAtPath)
+                .Select(importer => converter?.Invoke(importer) ?? importer as T)
+                .Where(source => source != null);
     }
 }